+++ packages/gateways/src/types/executor-gateway.ts
@@ -1,5 +1,9 @@
-import type {TransactionInfo} from 'fuzd-executor';
+import type {PendingExecutionStored} from 'fuzd-executor';
 
 export type ExecutorGateway = {
-	submitTransactionAsJsonString(id: string, submission: string, signature: `0x${string}`): Promise<TransactionInfo>;
+	submitTransactionAsJsonString(
+		id: string,
+		submission: string,
+		signature: `0x${string}`,
+	): Promise<PendingExecutionStored>;
 };